+++ test/data/user.ts
@@ -47,3 +47,54 @@ export const BOBATAN_TWISTED_MINDS_REALM_PERMISSIONS = [
   RealmPermissions.createRealmInvite,
   ...REALM_MEMBER_PERMISSIONS,
 ];
+
+// TODO: the realm_id for JERSEY_DEVIL_BOBADEX should not be v0, it should be the external_id of the realm (a UUID) - fix this when the relevant changes have been made elsewhere in the project (which will ideally make relevant tests fail)
+// TODO: identities_count for JERSEY_DEVIL_BOBADEX should be a number - fix this when changes have been made elsewhere (ideally making relevant tests fail)
+export const JERSEY_DEVIL_BOBADEX = {
+  seasons: [
+    {
+      id: "9f6d41a5-1e00-4071-9f50-555f1686f87f",
+      name: "Default",
+      realm_id: "v0",
+      identities_count: "3",
+      caught_identities: [
+        {
+          id: "85e33a3c-f987-41fd-a555-4c0cfdedf737",
+          name: "Old Time-y Anon",
+          index: 1,
+          avatar:
+            "https://firebasestorage.googleapis.com/v0/b/bobaboard-fb.appspot.com/o/images%2Fgore%2F5c2c3867-2323-4209-8bd4-9dfcc88808f3%2Fd931f284-5c22-422d-9343-e509cfb44ffc.png?alt=media&token=94e52fff-4e6b-4110-94c3-90b8800f541c",
+        },
+        {
+          id: "07f4cbbb-6a62-469e-8789-1b673a6d622f",
+          name: "DragonFucker",
+          index: 2,
+          avatar:
+            "https://pbs.twimg.com/profile_images/473496567366705152/JyHRKG7g.jpeg",
+        },
+      ],
+    },
+    {
+      id: "9b496931-ba27-43e0-953b-c38e01803879",
+      name: "Halloween",
+      realm_id: "v0",
+      identities_count: "3",
+      caught_identities: [
+        {
+          id: "47bf62fa-755f-489f-9735-27c884a0dec3",
+          name: "The OG OG Komaeda",
+          index: 3,
+          avatar:
+            "https://firebasestorage.googleapis.com/v0/b/bobaboard-fb.appspot.com/o/images%2Fbobaland%2Fundefined%2F1237fd9e-cd40-41b8-8ee7-11c865f27b6b?alt=media&token=4bb418a6-cb45-435c-85ed-7bdcb294f5b5",
+        },
+      ],
+    },
+    {
+      id: "be93274d-cdb9-4fcc-a4f9-a9c69270ce0d",
+      name: "Coders",
+      realm_id: "v0",
+      identities_count: "5",
+      caught_identities: [],
+    },
+  ],
+};
